requirements.txt
google-genai
python-dotenv>=1.0.0
tiktoken>=0.6.0
ollama>=0.1.5
numpy>=1.20.0  # Added for embedding similarity calculation
